Update on "Add deprecation warning for BlockSparseWeightConfig (#4230)"
BlockSparseWeightConfig is deprecated and will be removed in a future
release. A warning is now emitted on instantiation pointing users to
the tracking issue for more details.

Deprecation Note:
```
# 0.17:
BlockSparseWeightConfig(blocksize)

# 0.18 (deprecated):
BlockSparseWeightConfig(blocksize)

# 0.19 (removed):
See https://github.com/pytorch/ao/issues/4230 for more details.
```

[ghstack-poisoned]
